requirements.txt
asgiref==3.4.1
astroid==2.8.5
autopep8==1.6.0
belvo-python==0.29.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
Django==3.2.7
django-environ==0.9.0
django-filter==2.4.0
djangorestframework==3.12.4
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.2
isort==5.10.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
msgpack==1.0.4
platformdirs==2.4.0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.9
pytz==2021.1
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.6
wrapt==1.13.3
